Factored out GeneralAutoSendChecker and FormLevelAutoSendChecker to make testing easier
