+++ collect_app/src/main/java/org/odk/collect/android/backgroundwork/AutoSendTaskSpec.kt
@@ -14,8 +14,6 @@
 package org.odk.collect.android.backgroundwork
 
 import android.content.Context
-import android.net.ConnectivityManager
-import android.net.NetworkInfo
 import android.os.Environment
 import androidx.work.BackoffPolicy
 import androidx.work.WorkerParameters
@@ -25,11 +23,11 @@ import org.odk.collect.android.network.NetworkStateProvider
 import org.odk.collect.android.utilities.FormsRepositoryProvider
 import org.odk.collect.async.TaskSpec
 import org.odk.collect.async.WorkerAdapter
-import org.odk.collect.forms.Form
 import org.odk.collect.settings.SettingsProvider
-import org.odk.collect.settings.keys.ProjectKeys
 import java.util.function.Supplier
 import javax.inject.Inject
+import org.odk.collect.android.backgroundwork.autosend.FormLevelAutoSendChecker
+import org.odk.collect.android.backgroundwork.autosend.GeneralAutoSendChecker
 
 class AutoSendTaskSpec : TaskSpec {
     @Inject
@@ -43,6 +41,12 @@ class AutoSendTaskSpec : TaskSpec {
 
     @Inject
     lateinit var instanceAutoSender: InstanceAutoSender
+
+    @Inject
+    lateinit var generalAutoSendChecker: GeneralAutoSendChecker
+
+    @Inject
+    lateinit var formLevelAutoSendChecker: FormLevelAutoSendChecker
 
     override val maxRetries: Int? = null
     override val backoffPolicy: BackoffPolicy? = null
@@ -67,9 +71,8 @@ class AutoSendTaskSpec : TaskSpec {
         return Supplier {
             val projectId = inputData[TaskData.DATA_PROJECT_ID]
             if (projectId != null) {
-                val currentNetworkInfo = connectivityProvider.networkInfo
                 if (Environment.getExternalStorageState() == Environment.MEDIA_MOUNTED &&
-                    (networkTypeMatchesAutoSendSetting(currentNetworkInfo, projectId) || atLeastOneFormSpecifiesAutoSend(projectId))
+                    (generalAutoSendChecker.isAutoSendEnabled(projectId) || formLevelAutoSendChecker.isAutoSendEnabled(projectId))
                 ) {
                     return@Supplier instanceAutoSender.autoSendInstances(projectId)
                 }
@@ -82,41 +85,6 @@ class AutoSendTaskSpec : TaskSpec {
 
     override fun getWorkManagerAdapter(): Class<out WorkerAdapter> {
         return Adapter::class.java
-    }
-
-    /**
-     * Returns whether the currently-available connection type is included in the app-level auto-send
-     * settings.
-     *
-     * @return true if a connection is available and settings specify it should trigger auto-send,
-     * false otherwise.
-     */
-    private fun networkTypeMatchesAutoSendSetting(
-        currentNetworkInfo: NetworkInfo?,
-        projectId: String?
-    ): Boolean {
-        if (currentNetworkInfo == null) {
-            return false
-        }
-        val autosend = settingsProvider.getUnprotectedSettings(projectId).getString(ProjectKeys.KEY_AUTOSEND)
-        var sendwifi = autosend == "wifi_only"
-        var sendnetwork = autosend == "cellular_only"
-
-        if (autosend == "wifi_and_cellular") {
-            sendwifi = true
-            sendnetwork = true
-        }
-        return currentNetworkInfo.type == ConnectivityManager.TYPE_WIFI &&
-            sendwifi || currentNetworkInfo.type == ConnectivityManager.TYPE_MOBILE && sendnetwork
-    }
-
-    /**
-     * Returns true if at least one form currently on the device specifies that all of its filled
-     * forms should auto-send no matter the connection type.
-     */
-    private fun atLeastOneFormSpecifiesAutoSend(projectId: String?): Boolean {
-        return formsRepositoryProvider.get(projectId).all.stream()
-            .anyMatch { form: Form -> java.lang.Boolean.parseBoolean(form.autoSend) }
     }
 
     class Adapter(context: Context, workerParams: WorkerParameters) :
